+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.1.0
+
+### Features
+
+* Added support for blacklisted passwords
+* Added support for blacklisted strings within passwords
+
 ## v1.0.0
 
 Initial Release